@@ -2,43 +2,20 @@
   "name": "react-sandbox",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@dnd-kit/core": "^4.0.3",
-    "@dnd-kit/sortable": "^5.1.0",
-    "@dnd-kit/utilities": "^3.0.2",
-    "@nextui-org/react": "^1.0.0-beta.12",
-    "@textlint/kernel": "^13.3.1",
-    "@textlint/module-interop": "^13.3.1",
-    "@textlint/textlint-plugin-text": "^13.3.1",
-    "clsx": "^1.1.1",
-    "destyle.css": "^2.0.2",
-    "lodash-es": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.1",
-    "sass": "^1.45.1",
-    "styled-components": "^5.3.3",
-    "textlint-rule-preset-japanese": "^7.0.0",
-    "typescript": "^4.9.4",
-    "wasm": "file:./wasm/pkg",
-    "web-vitals": "^1.1.1"
+  "engines": {
+    "node": "20.x",
+    "pnpm": ">=3"
   },
   "scripts": {
     "postinstall": "typesync",
-    "dev": "yarn tsm:watch & yarn storybook",
     "start": "react-app-rewired start",
     "build": "react-app-rewired build",
     "eject": "react-app-rewired eject",
-    "fix": "yarn -s format && yarn -s lint:fix",
-    "format": "prettier --write --loglevel=warn '**/*.{js,jsx,ts,tsx,html,gql,graphql,json}'",
-    "lint": "yarn -s lint:style; yarn -s lint:es",
-    "lint:fix": "yarn -s lint:style:fix && yarn -s lint:es:fix",
+    "tsc": "tsc --noEmit",
     "lint:es": "eslint '**/*.{js,jsx,ts,tsx}'",
-    "lint:es:fix": "eslint --fix '**/*.{js,jsx,ts,tsx}'",
-    "lint:conflict": "eslint-config-prettier '**/*.{js,jsx,ts,tsx}'",
-    "lint:style": "stylelint '**/*.{css,scss}' 'src/components/**/styles.{ts,tsx}'",
-    "lint:style:fix": "stylelint --fix '**/*.{css,scss}' 'src/components/**/styles.{ts,tsx}'",
-    "tsc": "tsc",
+    "lint:style": "stylelint '**/*.{css,scss,ts,tsx}'",
+    "lint": "pnpm lint:es; pnpm lint:style",
+    "format": "prettier --write --loglevel=warn '**/*.{js,jsx,ts,tsx,html,gql,graphql,json}'",
     "hygen:component:new": "hygen component new",
     "tsm": "tsm src/components",
     "tsm:watch": "tsm src/components -w --ignoreInitial -u",
@@ -46,9 +23,99 @@
     "build-storybook": "storybook build -s public",
     "test": "jest",
     "test:update": "jest -u",
-    "code-check": "yarn lint & yarn tsc & yarn test:update",
     "cy:sb:open": "cypress open -C cypress.storybook.json",
-    "cy:sb:run": "cypress run --headless -C cypress.storybook.json"
+    "cy:sb:run": "cypress run --headless -C cypress.storybook.json",
+    "test:playwright": "playwright test --ui"
+  },
+  "dependencies": {
+    "@dnd-kit/core": "^6.1.0",
+    "@dnd-kit/sortable": "^8.0.0",
+    "@dnd-kit/utilities": "^3.2.2",
+    "@nextui-org/react": "^2.2.9",
+    "@suimenkathemove/frontend-tsconfig": "^0.1.7",
+    "@textlint/kernel": "^13.4.1",
+    "@textlint/module-interop": "^13.4.1",
+    "@textlint/textlint-plugin-text": "^13.4.1",
+    "clsx": "^2.1.0",
+    "destyle.css": "^4.0.0",
+    "lodash-es": "^4.17.21",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.69.7",
+    "styled-components": "^6.1.6",
+    "textlint-rule-preset-japanese": "^10.0.3",
+    "typescript": "5.2.2",
+    "wasm": "file:./wasm/pkg",
+    "web-vitals": "^3.5.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.23.7",
+    "@babel/preset-env": "^7.23.7",
+    "@babel/preset-typescript": "^7.23.3",
+    "@jest/types": "^29.6.3",
+    "@playwright/test": "^1.40.1",
+    "@storybook/addon-essentials": "^7.6.7",
+    "@storybook/addon-interactions": "^7.6.7",
+    "@storybook/addon-links": "^7.6.7",
+    "@storybook/jest": "^0.2.3",
+    "@storybook/node-logger": "^7.6.7",
+    "@storybook/react": "^7.6.7",
+    "@storybook/react-webpack5": "^7.6.7",
+    "@storybook/testing-library": "^0.2.2",
+    "@suimenkathemove/frontend-eslint-config": "^0.1.10",
+    "@suimenkathemove/prettier-config": "^0.1.4",
+    "@suimenkathemove/stylelint-config": "^0.1.4",
+    "@testing-library/jest-dom": "^6.2.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/babel__core": "^7.20.5",
+    "@types/babel__preset-env": "^7.9.6",
+    "@types/concurrently": "^6.4.0",
+    "@types/jest": "^29.5.11",
+    "@types/lodash-es": "^4.17.12",
+    "@types/node": "^20.10.6",
+    "@types/react": "^18.2.46",
+    "@types/react-dom": "^18.2.18",
+    "@types/sass-loader": "^8.0.8",
+    "@types/styled-components": "^5.1.34",
+    "@types/stylelint": "^13.13.3",
+    "@types/testing-library__jest-dom": "^5.14.9",
+    "@types/testing-library__user-event": "^4.2.0",
+    "babel-jest": "^29.7.0",
+    "change-case": "^5.3.0",
+    "concurrently": "^8.2.2",
+    "css-loader": "^6.8.1",
+    "cypress": "^13.6.2",
+    "hygen": "^6.2.11",
+    "jest": "^29.7.0",
+    "jest-css-modules": "^2.1.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-styled-components": "^7.2.0",
+    "mockdate": "^3.0.5",
+    "msw": "^2.0.11",
+    "react-app-rewired": "^2.2.1",
+    "reg-keygen-git-hash-plugin": "^0.12.2",
+    "reg-notify-github-plugin": "^0.12.2",
+    "reg-publish-s3-plugin": "^0.12.2",
+    "reg-suit": "^0.12.2",
+    "sass-loader": "^13.3.3",
+    "storybook": "^7.6.7",
+    "storycap": "^4.2.0",
+    "style-loader": "^3.3.3",
+    "stylelint": "^16.1.0",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.2",
+    "typed-scss-modules": "^8.0.0",
+    "typesync": "^0.12.0",
+    "wasm-loader": "^1.3.0",
+    "webpack": "^5.89.0"
+  },
+  "typesync": {
+    "ignorePackages": [
+      "wasm"
+    ]
   },
   "browserslist": {
     "production": [
