@@ -1,106 +1,18 @@
 module.exports = {
-  root: true,
-  env: {
-    browser: true,
-    es2021: true,
-    node: true,
-  },
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
-    },
-    ecmaVersion: 13,
-    sourceType: 'module',
-  },
-  plugins: [
-    'import',
-    '@typescript-eslint',
-    'react',
-    'react-hooks',
-    'jsx-a11y',
-    'testing-library',
-    'jest-dom',
-  ],
   extends: [
-    'eslint:recommended',
-    'plugin:import/errors',
-    'plugin:import/warnings',
-    'plugin:import/typescript',
-    'plugin:@typescript-eslint/recommended',
-    'plugin:react/recommended',
-    'plugin:react-hooks/recommended',
-    'plugin:jsx-a11y/recommended',
-    'plugin:testing-library/react',
-    'plugin:jest-dom/recommended',
-    'prettier',
+    // "@suimenkathemove/frontend-eslint-config",
+    './node_modules/@suimenkathemove/frontend-eslint-config',
   ],
-  rules: {
-    'padding-line-between-statements': [
-      'error',
-      {
-        blankLine: 'always',
-        prev: '*',
-        next: 'return',
-      },
-    ],
-    'import/no-named-as-default': 'off',
-    'import/no-unresolved': 'off',
-    'import/order': [
-      'error',
-      {
-        alphabetize: {
-          order: 'asc',
-        },
-        pathGroups: [
-          {
-            pattern: '@/**',
-            group: 'internal',
-          },
-        ],
-        pathGroupsExcludedImportTypes: ['builtin'],
-        'newlines-between': 'always',
-      },
-    ],
-    'no-console': 'error',
-    'sort-imports': [
-      'error',
-      {
-        ignoreDeclarationSort: true,
-      },
-    ],
-    '@typescript-eslint/ban-ts-comment': 'off',
-    '@typescript-eslint/no-empty-function': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/no-namespace': 'off',
-    '@typescript-eslint/no-non-null-assertion': 'off',
-    '@typescript-eslint/no-unused-vars': [
-      'error',
-      {
-        vars: 'all',
-        args: 'after-used',
-        argsIgnorePattern: '_',
-        ignoreRestSiblings: false,
-        varsIgnorePattern: '_',
-      },
-    ],
-    'react/display-name': 'off',
-    'react/react-in-jsx-scope': 'off',
-    'react-hooks/exhaustive-deps': 'error',
-    'react-hooks/rules-of-hooks': 'error',
-    'testing-library/no-wait-for-multiple-assertions': 'off',
-  },
   overrides: [
     {
-      files: ['*.tsx'],
+      files: ['cypress/**/*.ts', 'cypress/**/*.tsx'],
+      parserOptions: {
+        project: './cypress/tsconfig.json',
+      },
       rules: {
-        'react/prop-types': 'off',
+        '@typescript-eslint/no-floating-promises': 'error',
       },
     },
   ],
-  settings: {
-    react: {
-      version: 'detect',
-    },
-  },
+  ignorePatterns: ['wasm'],
 };
