@@ -1,23 +1,11 @@
 module.exports = {
-  plugins: ['stylelint-order'],
-  processors: [
-    [
-      'stylelint-processor-styled-components',
-      {
-        ignoreFiles: ['**/*.css', '**/*.scss'],
+  extends: ['@suimenkathemove/stylelint-config'],
+  overrides: [
+    {
+      files: ['**/*.module.scss'],
+      rules: {
+        'selector-class-pattern': null,
       },
-    ],
+    },
   ],
-  extends: [
-    'stylelint-config-standard',
-    'stylelint-config-recess-order',
-    'stylelint-config-styled-components',
-  ],
-  // MEMO: コメントアウトを外すとメッセージは表示されなくなるがlintが効かなくなる
-  // customSyntax: '@stylelint/postcss-css-in-js',
-  rules: {
-    'string-quotes': 'single',
-    'color-hex-length': 'long',
-  },
-  ignoreFiles: 'node_modules/**',
 };
