@@ -1,6 +1,6 @@
 import 'jest-styled-components';
 
-import { cleanup, render } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { defaultProps } from './index.stories';
 
@@ -9,10 +9,6 @@ import { SortableTreeSortableItem } from './';
 const setup = () => render(<SortableTreeSortableItem {...defaultProps} />);
 
 describe('SortableTreeSortableItem', () => {
-  afterEach(() => {
-    cleanup();
-  });
-
   test('render', () => {
     const { asFragment } = setup();
     expect(asFragment()).toMatchSnapshot();
