@@ -1,4 +1,4 @@
-import { cleanup, render } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { defaultProps } from './index.stories';
 
@@ -8,10 +8,6 @@ import { XButton } from './';
 const setup = () => render(<XButton {...defaultProps} />);
 
 describe('XButton', () => {
-  afterEach(() => {
-    cleanup();
-  });
-
   test('render', () => {
     const { asFragment } = setup();
     expect(asFragment()).toMatchSnapshot();
