@@ -5,7 +5,7 @@ import { wasmFlattenTree } from '../utils/wasm-flatten-tree';
 
 export const useTestWasmFlattenTree = (tree: Tree) => {
   useEffect(() => {
-    (async () => {
+    void (async () => {
       const flattenedTree = await wasmFlattenTree(tree);
       // eslint-disable-next-line no-console
       console.log(flattenedTree);
