@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import 'jest-styled-components';
 
-import { cleanup, render } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { defaultProps } from './index.stories';
 
@@ -10,10 +10,6 @@ import { SortableTree } from './';
 const setup = () => render(<SortableTree {...defaultProps} />);
 
 describe('SortableTree', () => {
-  afterEach(() => {
-    cleanup();
-  });
-
   test('render', () => {
     // const { asFragment } = setup();
     // expect(asFragment()).toMatchSnapshot();
