@@ -1,4 +1,4 @@
-import { cleanup, render } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { defaultProps } from './index.stories';
 
@@ -7,10 +7,6 @@ import { Grid } from './';
 const setup = () => render(<Grid {...defaultProps} />);
 
 describe('Grid', () => {
-  afterEach(() => {
-    cleanup();
-  });
-
   test('render', () => {
     const { asFragment } = setup();
     expect(asFragment()).toMatchSnapshot();
