@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 
 import styles from './styles.module.scss';
 
-import { Popover } from '@/components/ui-components/popover-old';
 import { MonthlyCalendar } from '@/components/ui-components/MonthlyCalendar';
+import { Popover } from '@/components/ui-components/popover-old';
 import { endOfDate } from '@/utils/date/end-of-date';
 import { isSameDate } from '@/utils/date/isSameDate';
 import { useShow } from '@/utils/useShow';
