@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { cleanup, render } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { defaultProps } from './index.stories';
 
@@ -9,10 +9,6 @@ import { Finder } from './';
 const setup = () => render(<Finder {...defaultProps} />);
 
 describe('Finder', () => {
-  afterEach(() => {
-    cleanup();
-  });
-
   test('render', () => {
     // const { asFragment } = setup();
     // expect(asFragment()).toMatchSnapshot();
