@@ -5,13 +5,12 @@ type ExtractPositionTypeUnit<T extends PositionTypeUnit> = T;
 type HorizontalPositionTypeUnit = ExtractPositionTypeUnit<'left' | 'right'>;
 type VerticalPositionTypeUnit = ExtractPositionTypeUnit<'top' | 'bottom'>;
 
-export type PositionTypeSecond<
-  PositionTypeFirst extends PositionTypeUnit
-> = PositionTypeFirst extends HorizontalPositionTypeUnit
-  ? VerticalPositionTypeUnit
-  : PositionTypeFirst extends VerticalPositionTypeUnit
-  ? HorizontalPositionTypeUnit
-  : never;
+export type PositionTypeSecond<PositionTypeFirst extends PositionTypeUnit> =
+  PositionTypeFirst extends HorizontalPositionTypeUnit
+    ? VerticalPositionTypeUnit
+    : PositionTypeFirst extends VerticalPositionTypeUnit
+      ? HorizontalPositionTypeUnit
+      : never;
 
 type CornerPositionType<T extends PositionTypeUnit> = T extends PositionTypeUnit
   ? `${T}-${PositionTypeSecond<T>}`
