@@ -1,5 +1,8 @@
 export class Position {
-  constructor(public x: number, public y: number) {
+  constructor(
+    public x: number,
+    public y: number,
+  ) {
     Object.assign(this, { x, y });
   }
 
