@@ -1,5 +1,8 @@
 export class Size {
-  constructor(public width: number, public height: number) {
+  constructor(
+    public width: number,
+    public height: number,
+  ) {
     Object.assign(this, { width, height });
   }
 
